+++ sys/netpfil/pf/pf_altq.h
@@ -29,7 +29,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  *	$OpenBSD: pfvar.h,v 1.282 2009/01/29 15:12:28 pyr Exp $
- *	$FreeBSD: 35d2d5cb8bbe75d64ace8fd38588b3797ef4f92a $
+ *	$FreeBSD: 0b913a19c2ff7a69755e9a2e5ea013f0aac65782 $
  */
 
 #ifndef	_NET_PF_ALTQ_H_
@@ -220,7 +220,7 @@ struct pf_kaltq {
 		struct fairq_opts        fairq_opts;
 	} pq_u;
 
-	uint32_t		 qid;		/* return value */
+	uint16_t		 qid;		/* return value */
 };
 #endif /* _KERNEL */
